+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Development
 - Use lxml.iterparse to reduce memory consumption when parsing DWD Mosmix files
 - Fix Settings object instantiation
 - Change logging level for Settings.cache_disable to INFO
+- Add DWD Observation climate_urban datasets
 
 0.42.1 (25.08.2022)
 *******************